Fix typo in package name: automaticfiededitor -> automaticfieldeditor
