--- jabgui/src/main/java/org/jabref/gui/edit/automaticfiededitor/AbstractAutomaticFieldEditorTabView.java
@@ -1,11 +1,0 @@
-package org.jabref.gui.edit.automaticfiededitor;
-
-import javafx.scene.layout.VBox;
-
-public abstract class AbstractAutomaticFieldEditorTabView extends VBox implements AutomaticFieldEditorTab {
-
-    @Override
-    public VBox getContent() {
-        return this;
-    }
-}
