--- jabgui/src/main/java/org/jabref/gui/edit/automaticfiededitor/AutomaticFieldEditorTab.java
@@ -1,9 +1,0 @@
-package org.jabref.gui.edit.automaticfiededitor;
-
-import javafx.scene.layout.Pane;
-
-public interface AutomaticFieldEditorTab {
-    Pane getContent();
-
-    String getTabName();
-}
