+++ jabgui/src/main/java/org/jabref/gui/edit/automaticfieldeditor/AutomaticFieldEditorTab.java
@@ -1,0 +1,9 @@
+package org.jabref.gui.edit.automaticfieldeditor;
+
+import javafx.scene.layout.Pane;
+
+public interface AutomaticFieldEditorTab {
+    Pane getContent();
+
+    String getTabName();
+}
