+++ jabgui/src/main/java/org/jabref/gui/edit/automaticfieldeditor/AutomaticFieldEditorUndoableEdit.java
@@ -1,0 +1,33 @@
+package org.jabref.gui.edit.automaticfieldeditor;
+
+import javax.swing.undo.CannotRedoException;
+import javax.swing.undo.CannotUndoException;
+
+import org.jabref.gui.undo.NamedCompoundEdit;
+
+public class AutomaticFieldEditorUndoableEdit extends NamedCompoundEdit {
+    int affectedEntries;
+
+    public AutomaticFieldEditorUndoableEdit(String name) {
+        super(name);
+        affectedEntries = 0;
+    }
+
+    public int getAffectedEntries() {
+        return affectedEntries;
+    }
+
+    public void setAffectedEntries(int affectedEntries) {
+        this.affectedEntries = affectedEntries;
+    }
+
+    @Override
+    public void undo() throws CannotUndoException {
+        super.undo();
+    }
+
+    @Override
+    public void redo() throws CannotRedoException {
+        super.redo();
+    }
+}
