+++ jabgui/src/main/java/org/jabref/gui/edit/automaticfieldeditor/copyormovecontent/CopyOrMoveFieldContentTabView.java
@@ -1,0 +1,114 @@
+package org.jabref.gui.edit.automaticfieldeditor.copyormovecontent;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import javafx.application.Platform;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.CheckBox;
+import javafx.scene.control.ComboBox;
+
+import org.jabref.gui.DialogService;
+import org.jabref.gui.StateManager;
+import org.jabref.gui.edit.automaticfieldeditor.AbstractAutomaticFieldEditorTabView;
+import org.jabref.gui.edit.automaticfieldeditor.AutomaticFieldEditorTab;
+import org.jabref.gui.edit.automaticfieldeditor.FieldHelper;
+import org.jabref.gui.undo.NamedCompoundEdit;
+import org.jabref.logic.l10n.Localization;
+import org.jabref.model.database.BibDatabase;
+import org.jabref.model.entry.BibEntry;
+import org.jabref.model.entry.field.Field;
+
+import com.airhacks.afterburner.views.ViewLoader;
+import com.tobiasdiez.easybind.EasyBind;
+import de.saxsys.mvvmfx.utils.validation.visualization.ControlsFxVisualizer;
+
+import static org.jabref.gui.util.FieldsUtil.FIELD_STRING_CONVERTER;
+
+public class CopyOrMoveFieldContentTabView extends AbstractAutomaticFieldEditorTabView implements AutomaticFieldEditorTab {
+    public Button copyContentButton;
+    private final NamedCompoundEdit compoundEdit;
+    private final DialogService dialogService;
+    private final List<BibEntry> selectedEntries;
+    private final BibDatabase database;
+    private final StateManager stateManager;
+    private final ControlsFxVisualizer visualizer = new ControlsFxVisualizer();
+    @FXML
+    private Button moveContentButton;
+    @FXML
+    private Button swapContentButton;
+    @FXML
+    private ComboBox<Field> fromFieldComboBox;
+    @FXML
+    private ComboBox<Field> toFieldComboBox;
+    @FXML
+    private CheckBox overwriteFieldContentCheckBox;
+    private CopyOrMoveFieldContentTabViewModel viewModel;
+
+    public CopyOrMoveFieldContentTabView(BibDatabase database,
+                                         NamedCompoundEdit compoundEdit,
+                                         DialogService dialogService,
+                                         StateManager stateManager) {
+        this.compoundEdit = compoundEdit;
+        this.dialogService = dialogService;
+        this.selectedEntries = new ArrayList<>(stateManager.getSelectedEntries());
+        this.database = database;
+        this.stateManager = stateManager;
+
+        ViewLoader.view(this)
+                  .root(this)
+                  .load();
+    }
+
+    public void initialize() {
+        viewModel = new CopyOrMoveFieldContentTabViewModel(database, selectedEntries, compoundEdit, dialogService, stateManager);
+        initializeFromAndToComboBox();
+
+        viewModel.overwriteFieldContentProperty().bindBidirectional(overwriteFieldContentCheckBox.selectedProperty());
+
+        moveContentButton.disableProperty().bind(viewModel.canMoveProperty().not());
+        swapContentButton.disableProperty().bind(viewModel.canSwapProperty().not());
+        copyContentButton.disableProperty().bind(viewModel.toFieldValidationStatus().validProperty().not());
+        overwriteFieldContentCheckBox.disableProperty().bind(viewModel.toFieldValidationStatus().validProperty().not());
+
+        Platform.runLater(() -> visualizer.initVisualization(viewModel.toFieldValidationStatus(), toFieldComboBox, true));
+    }
+
+    private void initializeFromAndToComboBox() {
+        // From
+        fromFieldComboBox.getItems().setAll(FieldHelper.getSetFieldsOnly(selectedEntries, viewModel.getAllFields()));
+        fromFieldComboBox.setConverter(FIELD_STRING_CONVERTER);
+        fromFieldComboBox.valueProperty().bindBidirectional(viewModel.fromFieldProperty());
+        EasyBind.listen(fromFieldComboBox.getEditor().textProperty(), _ -> fromFieldComboBox.commitValue());
+        if (!fromFieldComboBox.getItems().isEmpty()) {
+            fromFieldComboBox.getSelectionModel().selectFirst();
+        }
+
+        // To
+        toFieldComboBox.getItems().setAll(viewModel.getAllFields());
+        toFieldComboBox.setConverter(FIELD_STRING_CONVERTER);
+        toFieldComboBox.valueProperty().bindBidirectional(viewModel.toFieldProperty());
+        EasyBind.listen(toFieldComboBox.getEditor().textProperty(), _ -> toFieldComboBox.commitValue());
+    }
+
+    @Override
+    public String getTabName() {
+        return Localization.lang("Copy or move content");
+    }
+
+    @FXML
+    void copyContent() {
+        viewModel.copyValue();
+    }
+
+    @FXML
+    void moveContent() {
+        viewModel.moveValue();
+    }
+
+    @FXML
+    void swapContent() {
+        viewModel.swapValues();
+    }
+}
