--- jabgui/src/main/resources/org/jabref/gui/edit/automaticfiededitor/AutomaticFieldEditorDialog.fxml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.ButtonType?>
-<?import javafx.scene.control.DialogPane?>
-<?import javafx.scene.control.TabPane?>
-<DialogPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.jabref.gui.edit.automaticfiededitor.AutomaticFieldEditorDialog"
-            prefHeight="400.0" prefWidth="600.0" styleClass="automatic-field-editor">
-    <content>
-        <TabPane fx:id="tabPane" tabClosingPolicy="UNAVAILABLE"/>
-    </content>
-    <ButtonType text="%Keep modifications" buttonData="OK_DONE" />
-    <ButtonType fx:constant="CANCEL"/>
-</DialogPane>
