--- jabgui/src/main/resources/org/jabref/gui/edit/automaticfiededitor/clearcontent/ClearContentTab.fxml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.layout.VBox?>
-<?import org.jabref.gui.icon.JabRefIconView?>
-<fx:root
-        prefHeight="400.0"
-        prefWidth="600.0"
-        type="VBox"
-        styleClass="edit-field-content-pane"
-        xmlns="http://javafx.com/javafx/17"
-        xmlns:fx="http://javafx.com/fxml/1"
-        fx:controller="org.jabref.gui.edit.automaticfiededitor.clearcontent.ClearContentTabView">
-    <GridPane
-            hgap="8.0"
-            prefWidth="568.0"
-            vgap="8.0">
-        <columnConstraints>
-            <ColumnConstraints
-                    hgrow="SOMETIMES"
-                    minWidth="10.0"
-                    prefWidth="100.0"/>
-            <ColumnConstraints
-                    hgrow="SOMETIMES"
-                    minWidth="10.0"
-                    prefWidth="100.0"/>
-            <ColumnConstraints
-                    hgrow="SOMETIMES"
-                    minWidth="10.0"
-                    prefWidth="100.0"/>
-        </columnConstraints>
-        <rowConstraints>
-            <RowConstraints
-                    minHeight="10.0"
-                    prefHeight="30.0"
-                    vgrow="SOMETIMES"/>
-            <RowConstraints
-                    minHeight="10.0"
-                    prefHeight="30.0"
-                    vgrow="SOMETIMES"/>
-            <RowConstraints
-                    minHeight="10.0"
-                    prefHeight="30.0"
-                    vgrow="SOMETIMES"/>
-        </rowConstraints>
-        <HBox alignment="BOTTOM_LEFT"
-              maxHeight="1.7976931348623157E308"
-              maxWidth="1.7976931348623157E308"
-              minHeight="-Infinity"
-              minWidth="-Infinity"
-              GridPane.hgrow="ALWAYS"
-              GridPane.rowSpan="2"
-              GridPane.vgrow="ALWAYS">
-            <Label styleClass="sectionHeader"
-                   text="%Clear field content for selected entries">
-                <HBox.margin>
-                    <Insets bottom="10.0"/>
-                </HBox.margin>
-            </Label>
-        </HBox>
-        <HBox spacing="8.0"
-              GridPane.rowIndex="2"
-              GridPane.columnSpan="3"
-              GridPane.hgrow="ALWAYS">
-            <ComboBox
-                    fx:id="fieldComboBox"
-                    prefHeight="36.0"
-                    editable="true"
-                    HBox.hgrow="ALWAYS"/>
-        </HBox>
-        <HBox spacing="8.0"
-              GridPane.rowIndex="3"
-              GridPane.columnSpan="3">
-            <Button fx:id="clearButton"
-                    text="%Clear field content"
-                    onAction="#onClear">
-                <graphic>
-                    <JabRefIconView
-                            glyph="DELETE_ENTRY"/>
-                </graphic>
-            </Button>
-        </HBox>
-    </GridPane>
-</fx:root>
