--- jabgui/src/main/resources/org/jabref/gui/edit/automaticfiededitor/copyormovecontent/CopyOrMoveFieldContentTab.fxml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.layout.VBox?>
-<fx:root
-        type="VBox"
-        styleClass="edit-field-content-pane"
-        xmlns="http://javafx.com/javafx/17"
-        xmlns:fx="http://javafx.com/fxml/1"
-        fx:controller="org.jabref.gui.edit.automaticfiededitor.copyormovecontent.CopyOrMoveFieldContentTabView">
-    <Label styleClass="sectionHeader"
-           text="%Copy or move the content of one field to another"/>
-    <GridPane
-            styleClass="grid-pane">
-        <columnConstraints>
-            <ColumnConstraints
-                    hgrow="NEVER"/>
-            <ColumnConstraints
-                    hgrow="ALWAYS"/>
-        </columnConstraints>
-        <rowConstraints>
-            <RowConstraints/>
-            <RowConstraints/>
-        </rowConstraints>
-        <Label text="%From"
-               GridPane.columnIndex="0"
-               GridPane.rowIndex="0"/>
-        <ComboBox
-                fx:id="fromFieldComboBox"
-                editable="true"
-                maxWidth="Infinity"
-                GridPane.columnIndex="1"
-                GridPane.rowIndex="0"/>
-        <Label text="%To"
-               GridPane.columnIndex="0"
-               GridPane.rowIndex="1"/>
-        <ComboBox
-                fx:id="toFieldComboBox"
-                editable="true"
-                maxWidth="Infinity"
-                GridPane.columnIndex="1"
-                GridPane.rowIndex="1"/>
-    </GridPane>
-    <CheckBox
-            fx:id="overwriteFieldContentCheckBox"
-            mnemonicParsing="false"
-            text="%Overwrite field content"/>
-    <HBox styleClass="actions">
-        <Button fx:id="copyContentButton"
-                mnemonicParsing="false"
-                onAction="#copyContent"
-                text="%Copy content"/>
-        <Button fx:id="moveContentButton"
-                mnemonicParsing="false"
-                onAction="#moveContent"
-                text="%Move content"/>
-        <Button fx:id="swapContentButton"
-                mnemonicParsing="false"
-                onAction="#swapContent"
-                text="%Swap content"/>
-    </HBox>
-</fx:root>
