--- jabgui/src/main/resources/org/jabref/gui/edit/automaticfiededitor/renamefield/RenameFieldTab.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<fx:root
-        type="VBox"
-        styleClass="edit-field-content-pane"
-        xmlns="http://javafx.com/javafx/17"
-        xmlns:fx="http://javafx.com/fxml/1"
-        fx:controller="org.jabref.gui.edit.automaticfiededitor.renamefield.RenameFieldTabView">
-    <Label styleClass="sectionHeader"
-           text="%Rename field"/>
-    <HBox styleClass="actions">
-        <ComboBox
-                fx:id="fieldComboBox"
-                editable="true"
-                HBox.hgrow="ALWAYS"/>
-        <TextField
-                fx:id="newFieldNameTextField"
-                promptText="%New field name"
-                HBox.hgrow="ALWAYS"/>
-    </HBox>
-    <Button fx:id="renameButton"
-            mnemonicParsing="false"
-            onAction="#renameField"
-            text="%Rename"/>
-</fx:root>
