+++ jabgui/src/main/resources/org/jabref/gui/edit/automaticfieldeditor/editfieldcontent/EditFieldContentTab.fxml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.Separator?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<fx:root
+        type="VBox"
+        styleClass="edit-field-content-pane"
+        xmlns="http://javafx.com/javafx/17"
+        xmlns:fx="http://javafx.com/fxml/1"
+        fx:controller="org.jabref.gui.edit.automaticfieldeditor.editfieldcontent.EditFieldContentTabView">
+    <Label styleClass="sectionHeader"
+           text="%Edit field content for selected entries"/>
+    <HBox styleClass="actions">
+        <ComboBox
+                fx:id="fieldComboBox"
+                editable="true"
+                HBox.hgrow="ALWAYS"/>
+        <TextField
+                fx:id="fieldValueTextField"
+                promptText="%Set or append content"
+                HBox.hgrow="ALWAYS"/>
+    </HBox>
+    <CheckBox
+            fx:id="showOnlySetFieldsCheckBox"
+            mnemonicParsing="false"
+            text="%Show only set fields"/>
+
+    <CheckBox
+            fx:id="overwriteFieldContentCheckBox"
+            mnemonicParsing="false"
+            text="%Overwrite field content"/>
+    <HBox styleClass="actions">
+        <Button fx:id="setValueButton"
+                mnemonicParsing="false"
+                onAction="#setFieldValue"
+                text="%Set"/>
+        <Button fx:id="appendValueButton"
+                mnemonicParsing="false"
+                onAction="#appendToFieldValue"
+                text="%Append"/>
+        <Separator
+                orientation="VERTICAL"/>
+    </HBox>
+</fx:root>
